pyproject.toml
[project]
name = "aws-config-to-s3"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = ["aws-cdk-lib>=2.189.0"]

[dependency-groups]
dev = ["boto3>=1.37.31", "pandas>=2.2.3"]
